# apps/whatsapp-api/docs/api-documentation.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Next Mavens Fidscript WhatsApp API",
    "description": "Production-ready WhatsApp Business API with anti-ban protection, rate limiting, and latest WhatsApp features. Multi-tenant SaaS platform for sending and receiving WhatsApp messages programmatically.",
    "version": "1.0.0",
    "contact": {
      "name": "Next Mavens Support",
      "email": "support@nextmavens.com"
    },
    "license": {
      "name": "Commercial",
      "url": "https://nextmavens.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development"
    }
  ],
...
  "paths": {
...
    "/anti-ban/rate-limit/{instanceName}/{phone}": {
      "get": {
        "tags": ["Anti-Ban"],
        "summary": "Check rate limit status for a contact",
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "instanceName", "in": "path", "required": true, "schema": {"type": "string"}},
          {"name": "phone", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {
            "description": "Rate limit status retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "allowed": {"type": "boolean"},
                    "reason": {"type": "string"},
                    "retryAfterMs": {"type": "integer"},
                    "currentCount": {"type": "integer"},
                    "limit": {"type": "integer"}
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/anti-ban/contact/{instanceName}/{phone}": {
      "get": {
        "tags": ["Anti-Ban"],
        "summary": "Get contact status (blocked/suppressed)",
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "instanceName", "in": "path", "required": true, "schema": {"type": "string"}},
          {"name": "phone", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {
            "description": "Contact status retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "phoneNumber": {"type": "string"},
                    "isSuppressed": {"type": "boolean"},
                    "totalBlocks": {"type": "integer"},
                    "suppressionReason": {"type": "string"},
                    "ttlRemaining": {"type": "integer"}
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/analytics/instance/{instanceName}": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get analytics for a specific instance",
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "instanceName", "in": "path", "required": true, "schema": {"type": "string"}},
          {"name": "startDate", "in": "query", "schema": {"type": "string", "format": "date-time"}},
          {"name": "endDate", "in": "query", "schema": {"type": "string", "format": "date-time"}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 1000}}
        ],
        "responses": {
          "200": {
            "description": "Instance analytics retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "instanceName": {"type": "string"},
                    "status": {"type": "string"},
                    "uptime": {"type": "integer"},
                    "messageStats": {
                      "type": "object",
                      "properties": {
                        "total": {"type": "integer"},
                        "sent": {"type": "integer"},
                        "delivered": {"type": "integer"},
                        "read": {"type": "integer"},
                        "failed": {"type": "integer"},
                        "pending": {"type": "integer"}
                      }
                    },
                    "deliveryRate": {
                      "type": "object",
                      "properties": {
                        "deliveryRate": {"type": "number"},
                        "readRate": {"type": "number"},
                        "failedRate": {"type": "number"}
                      }
                    },
                    "topContacts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "phone": {"type": "string"},
                          "messageCount": {"type": "integer"}
                        }
                      }
                    },
                    "hourlyStats": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "hour": {"type": "integer"},
                          "total": {"type": "integer"},
                          "sent": {"type": "integer"},
                          "delivered": {"type": "integer"},
                          "read": {"type": "integer"}
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/analytics/platform": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get platform-wide analytics",
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "startDate", "in": "query", "schema": {"type": "string", "format": "date-time"}},
          {"name": "endDate", "in": "query", "schema": {"type": "string", "format": "date-time"}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 1000}}
        ],
        "responses": {
          "200": {
            "description": "Platform analytics retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalInstances": {"type": "integer"},
                    "connectedInstances": {"type": "integer"},
                    "disconnectedInstances": {"type": "integer"},
                    "totalMessages": {"type": "integer"},
                    "totalDelivered": {"type": "integer"},
                    "totalRead": {"type": "integer"},
                    "totalFailed": {"type": "integer"},
                    "overallDeliveryRate": {"type": "number"},
                    "overallReadRate": {"type": "number"},
                    "topInstances": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "instanceName": {"type": "string"},
                          "messageCount": {"type": "integer"}
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/analytics/realtime/{instanceName}": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get real-time metrics for an instance",
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "instanceName", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {
            "description": "Real-time metrics retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "messagesLastHour": {"type": "integer"},
                    "deliveryRateLastHour": {"type": "number"},
                    "activeContacts": {"type": "integer"}
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "apikey",
        "description": "API key authentication"
      }
    },
...
  }
}